Add CreateNodePool and ScaleOut actions for Cassandra.
* These become the only changes supported by the Cassandra controller.
* ScaleIn and CassandraUpgrade actions will be implemented in followup branches.

Fixes: #253

+++ pkg/apis/navigator/types.go
@@ -50,7 +50,7 @@ type CassandraClusterStatus struct {
 }
 
 type CassandraClusterNodePoolStatus struct {
-	ReadyReplicas int64
+	ReadyReplicas int32
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
